user_service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.1",
    "fastapi[standard]>=0.115.11",
    "jwt>=1.3.1",
    "passlib>=1.7.4",
    "psycopg[binary]>=3.2.5",
    "pydantic-settings>=2.8.1",
    "sqlmodel>=0.0.23",
    "starlette>=0.46.0",
]
